LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-event-handled.https-expected.txt

CONSOLE MESSAGE: Fetch event was canceled
CONSOLE MESSAGE: Fetch API cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample.txt?respondWith-not-called-and-event-canceled.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: Returned response is null.
CONSOLE MESSAGE: Fetch API cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample.txt?respondWith-called-and-promise-resolved-to-invalid-response.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: Error: respondWith rejected
CONSOLE MESSAGE: Fetch API cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample.txt?respondWith-called-and-promise-rejected.

PASS global setup
PASS FetchEvent.handled should resolve when respondWith() is not called for a navigation request
PASS FetchEvent.handled should resolve when respondWith() is not called for a sub-resource request
PASS FetchEvent.handled should reject when respondWith() is not called and the event is canceled
PASS FetchEvent.handled should resolve when the promise provided to respondWith() is resolved
PASS FetchEvent.handled should reject when the promise provided to respondWith() is resolved to an invalid response
PASS FetchEvent.handled should reject when the promise provided to respondWith() is rejected
PASS global cleanup
